Refactor wishlist partials
