+++ web/app/styles/themes/_pw-components.scss
@@ -30,6 +30,7 @@
 @import 'node_modules/progressive-web-sdk/dist/components/card-verification/base';
 @import 'node_modules/progressive-web-sdk/dist/components/carousel/base';
 @import 'node_modules/progressive-web-sdk/dist/components/divider/base';
+@import 'node_modules/progressive-web-sdk/dist/components/feedback/base';
 @import 'node_modules/progressive-web-sdk/dist/components/field/base';
 @import 'node_modules/progressive-web-sdk/dist/components/field-row/base';
 @import 'node_modules/progressive-web-sdk/dist/components/grid/base';
@@ -83,6 +84,7 @@
 @import 'pw-components/card-verification';
 @import 'pw-components/carousel';
 @import 'pw-components/divider';
+@import 'pw-components/feedback';
 @import 'pw-components/field';
 @import 'pw-components/field-row';
 @import 'pw-components/header-bar';
