+++ web/app/styles/themes/pw-components/_feedback.scss
@@ -1,0 +1,24 @@
+// Feedback Theme
+// ===
+
+
+// Themes: Error
+// ---
+
+.pw-feedback.pw--error {
+    color: $error-color;
+}
+
+
+
+// Themes: Block
+// ---
+
+.pw-feedback.pw--block {
+    min-height: $tap-size;
+    padding: $unit*2;
+
+    &.pw--error {
+        background: $light-error-color;
+    }
+}
